Move exported errors to a single file
