+++ marketdata/stream/client.go
@@ -246,9 +246,6 @@ func (cc *cryptoClient) constructURL() (url.URL, error) {
 	return url.URL{Scheme: scheme, Host: ub.Host, Path: ub.Path, RawQuery: rawQuery}, nil
 }
 
-// ErrConnectCalledMultipleTimes is returned when Connect has been called multiple times on a single client
-var ErrConnectCalledMultipleTimes = errors.New("tried to call Connect multiple times")
-
 func (c *client) connect(ctx context.Context, u url.URL) error {
 	err := ErrConnectCalledMultipleTimes
 	c.connectOnce.Do(func() {
