+++ marketdata/stream/flow.go
@@ -81,10 +81,6 @@ func (c *client) initialize(ctx context.Context) error {
 	return nil
 }
 
-// ErrNoConnected is returned when the client did not receive the welcome
-// message from the server
-var ErrNoConnected = errors.New("did not receive connected message")
-
 func (c *client) readConnected(ctx context.Context) error {
 	b, err := c.conn.readMessage(ctx)
 	if err != nil {
@@ -118,33 +114,6 @@ func (c *client) writeAuth(ctx context.Context) error {
 
 	return c.conn.writeMessage(ctx, msg)
 }
-
-//ErrBadAuthResponse is returned when the client could not successfully authenticate
-var ErrBadAuthResponse = errors.New("did not receive authenticated message")
-
-// ErrConnectionLimitExceeded is returned when the client has exceeded their connection limit
-var ErrConnectionLimitExceeded error = errorMessage{msg: "connection limit exceeded", code: 406}
-
-// ErrInvalidCredentials is returned when invalid credentials have been sent by the user.
-var ErrInvalidCredentials error = errorMessage{msg: "auth failed", code: 402}
-
-// ErrInsufficientSubscription is returned when the user does not have proper
-// subscription for the requested feed (e.g. SIP)
-var ErrInsufficientSubscription error = errorMessage{msg: "insufficient subscription", code: 409}
-
-// ErrInsufficientScope is returned when the token used by the user doesn't have proper scopes
-// for data stream
-var ErrInsufficientScope error = errorMessage{msg: "insufficient scope", code: 411}
-
-// ErrSymbolLimitExceeded is returned when the client has subscribed to too many symbols
-var ErrSymbolLimitExceeded error = errorMessage{msg: "symbol limit exceeded", code: 405}
-
-// ErrSlowClient is returned when the server has detected a slow client. In this case there's no guarantee
-// that all prior messages are sent to the server so a subscription acknowledgement may not arrive
-var ErrSlowClient error = errorMessage{msg: "slow client", code: 407}
-
-// ErrSubscriptionChangeTimeout is returned when a subscription change is invalid for the feed.
-var ErrSubscriptionChangeInvalidForFeed error = errorMessage{msg: "invalid subscribe action for this feed", code: 410}
 
 // isErrorRetriable returns whether the error is considered retriable during the initialization flow
 func isErrorRetriable(err error) bool {
@@ -191,10 +160,6 @@ func (c *client) writeSub(ctx context.Context) error {
 
 	return c.conn.writeMessage(ctx, msg)
 }
-
-// ErrSubResponse is returned when the client's subscription request was not
-// acknowledged
-var ErrSubResponse = errors.New("did not receive subscribed message")
 
 func (c *client) readSubResponse(ctx context.Context) error {
 	b, err := c.conn.readMessage(ctx)
